+++ test/Unit/Reporter/Console/ConsoleReporterTest.php
@@ -17,12 +17,14 @@ use Ergebnis\PHPUnit\SlowTestDetector\Count;
 use Ergebnis\PHPUnit\SlowTestDetector\Duration;
 use Ergebnis\PHPUnit\SlowTestDetector\MaximumCount;
 use Ergebnis\PHPUnit\SlowTestDetector\MaximumDuration;
+use Ergebnis\PHPUnit\SlowTestDetector\MaximumWidth;
 use Ergebnis\PHPUnit\SlowTestDetector\Reporter;
 use Ergebnis\PHPUnit\SlowTestDetector\SlowTest;
 use Ergebnis\PHPUnit\SlowTestDetector\SlowTestList;
 use Ergebnis\PHPUnit\SlowTestDetector\Test;
 use Ergebnis\PHPUnit\SlowTestDetector\TestDescription;
 use Ergebnis\PHPUnit\SlowTestDetector\TestIdentifier;
+use Ergebnis\PHPUnit\SlowTestDetector\Width;
 use PHPUnit\Framework;
 
 /**
@@ -33,12 +35,14 @@ use PHPUnit\Framework;
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\Duration
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\MaximumCount
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\MaximumDuration
+ * @uses \Ergebnis\PHPUnit\SlowTestDetector\MaximumWidth
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\Reporter\Console\DurationFormatter
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\Reporter\Console\Unit
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\SlowTest
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\SlowTestList
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\TestDescription
  * @uses \Ergebnis\PHPUnit\SlowTestDetector\TestIdentifier
+ * @uses \Ergebnis\PHPUnit\SlowTestDetector\Width
  */
 final class ConsoleReporterTest extends Framework\TestCase
 {
@@ -53,7 +57,8 @@ final class ConsoleReporterTest extends Framework\TestCase
         $reporter = new Reporter\Console\ConsoleReporter(
             new Reporter\Console\DurationFormatter(),
             MaximumDuration::default(),
-            MaximumCount::fromCount(Count::fromInt($faker->numberBetween(1)))
+            MaximumCount::fromCount(Count::fromInt($faker->numberBetween(1))),
+            MaximumWidth::unlimited()
         );
 
         $report = $reporter->report($slowTestList);
@@ -73,7 +78,8 @@ final class ConsoleReporterTest extends Framework\TestCase
         $reporter = new Reporter\Console\ConsoleReporter(
             new Reporter\Console\DurationFormatter(),
             $maximumDuration,
-            $maximumCount
+            $maximumCount,
+            MaximumWidth::unlimited()
         );
 
         $report = $reporter->report($slowTestList);
@@ -525,6 +531,138 @@ final class ConsoleReporterTest extends Framework\TestCase
                 $expected,
                 $maximumDuration,
                 $maximumCount,
+                $slowTestList,
+            ];
+        }
+    }
+
+    /**
+     * @dataProvider provideExpectedReportMaximumWidthAndSlowTestList
+     */
+    public function testReportReturnsReportWithTestDescriptionsTruncatedToMaximumWidth(
+        string $expectedReport,
+        MaximumWidth $maximumWidth,
+        SlowTestList $slowTestList
+    ) {
+        $reporter = new Reporter\Console\ConsoleReporter(
+            new Reporter\Console\DurationFormatter(),
+            MaximumDuration::fromDuration(Duration::fromMilliseconds(100)),
+            MaximumCount::fromCount(Count::fromInt(1)),
+            $maximumWidth
+        );
+
+        $report = $reporter->report($slowTestList);
+
+        self::assertSame($expectedReport, $report);
+    }
+
+    /**
+     * @return \Generator<string, array{0: string, 1: MaximumWidth, 2: SlowTestList}>
+     */
+    public static function provideExpectedReportMaximumWidthAndSlowTestList(): iterable
+    {
+        $print = static function (array $lines): string {
+            return \implode(
+                "\n",
+                $lines
+            );
+        };
+
+        $testDescription = 'FooTest::testSleeperSleepsLongerThanMaximumDuration with data set #0 (1000)';
+
+        $slowTestWithGlobalMaximumDuration = SlowTest::create(
+            TestIdentifier::fromString('FooTest::test'),
+            TestDescription::fromString($testDescription),
+            Duration::fromMilliseconds(300),
+            MaximumDuration::fromDuration(Duration::fromMilliseconds(100))
+        );
+
+        $slowTestWithCustomMaximumDuration = SlowTest::create(
+            TestIdentifier::fromString('FooTest::test'),
+            TestDescription::fromString($testDescription),
+            Duration::fromMilliseconds(300),
+            MaximumDuration::fromDuration(Duration::fromMilliseconds(200))
+        );
+
+        $slowTestWithMultibyteTestDescription = SlowTest::create(
+            TestIdentifier::fromString('FooTest::test'),
+            TestDescription::fromString('FooTest::testWithÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜ'),
+            Duration::fromMilliseconds(300),
+            MaximumDuration::fromDuration(Duration::fromMilliseconds(100))
+        );
+
+        $values = [
+            'global-only-truncated' => [
+                $print([
+                    '',
+                    '',
+                    'Detected 1 test where the duration exceeded the global maximum duration (0.100).',
+                    '',
+                    '# Duration Test',
+                    '--------------------------------------------------------------------------------',
+                    '1    0.300 FooTest::testSleeperSleepsLongerTh…umDuration with data set #0 (1000)',
+                    '--------------------------------------------------------------------------------',
+                    '     0.000',
+                    '      └─── seconds',
+                ]),
+                MaximumWidth::fromWidth(Width::fromInt(80)),
+                SlowTestList::create($slowTestWithGlobalMaximumDuration),
+            ],
+            'custom-and-global-truncated' => [
+                $print([
+                    '',
+                    '',
+                    'Detected 1 test where the duration exceeded a custom or the global maximum duration (0.100).',
+                    '',
+                    '# Duration          Test',
+                    '  Actual   Maximum',
+                    '--------------------------------------------------------------------------------',
+                    '1    0.300    0.200 FooTest::testSleeperSleepsLong…ation with data set #0 (1000)',
+                    '--------------------------------------------------------------------------------',
+                    '     0.000',
+                    '      └─── seconds',
+                ]),
+                MaximumWidth::fromWidth(Width::fromInt(80)),
+                SlowTestList::create($slowTestWithCustomMaximumDuration),
+            ],
+            'greater-than-required' => [
+                $print([
+                    '',
+                    '',
+                    'Detected 1 test where the duration exceeded the global maximum duration (0.100).',
+                    '',
+                    '# Duration Test',
+                    '--------------------------------------------------------------------------------------',
+                    '1    0.300 FooTest::testSleeperSleepsLongerThanMaximumDuration with data set #0 (1000)',
+                    '--------------------------------------------------------------------------------------',
+                    '     0.000',
+                    '      └─── seconds',
+                ]),
+                MaximumWidth::fromWidth(Width::fromInt(200)),
+                SlowTestList::create($slowTestWithGlobalMaximumDuration),
+            ],
+            'multibyte-test-description' => [
+                $print([
+                    '',
+                    '',
+                    'Detected 1 test where the duration exceeded the global maximum duration (0.100).',
+                    '',
+                    '# Duration Test',
+                    '--------------------------------------------------------------------------------',
+                    '1    0.300 FooTest::testWithÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖ…ÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜÄÖÜ',
+                    '--------------------------------------------------------------------------------',
+                    '     0.000',
+                    '      └─── seconds',
+                ]),
+                MaximumWidth::fromWidth(Width::fromInt(80)),
+                SlowTestList::create($slowTestWithMultibyteTestDescription),
+            ],
+        ];
+
+        foreach ($values as $key => list($expected, $maximumWidth, $slowTestList)) {
+            yield $key => [
+                $expected,
+                $maximumWidth,
                 $slowTestList,
             ];
         }